requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
requests==2.26.0
faiss-cpu==1.7.2
numpy==1.21.2
pandas==1.3.3 